requirements.txt
requests>=2.25.0
pyjwt>=2.0.0
